extract selector context to concerned class
